Feat/message api setting

See merge request s11-webmobile4-sub2/S11P12D211!60

+++ android/Signal/app/src/main/java/com/ongo/signal/data/repository/chat/ChatRoomRepositoryImpl.kt
@@ -16,7 +16,7 @@ class ChatRoomRepositoryImpl @Inject constructor(
         chatHomeDao.insertAll(room)
     }
 
-    suspend fun getAllMessages(ID: Int): List<ChatHomeChildDto> = chatDetailDao.getAll(ID)
+    suspend fun getAllMessages(ID: Long): List<ChatHomeChildDto> = chatDetailDao.getAll(ID)
 
     suspend fun insertMessage(message: ChatHomeChildDto) {
         chatDetailDao.insertMessage(message)
